requirements.txt
astor==0.8.1
cycler==0.10.0
keras==2.4.3
google-pasta==0.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
numpy==1.16.4
opt-einsum==3.2.0
Pillow==6.0.0
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
scipy==1.2.1
tensorboard==1.14.0
tensorflow==2.2.0
tensorflow-estimator==1.14.0
torch==1.4.0
torchvision==0.5.0


